Update exercice_1-3_calcul-de-la-moyenne
Modification de paramètres de variables

+++ DOSSIER_03_ALGORITHMIQUE/exercice_1-3_calcul-de-la-moyenne/exercice_1-3_calcul-de-la-moyenne/Program.cs
@@ -8,10 +8,10 @@ namespace exercice_1_3_calcul_de_la_moyenne
 
         {
             Console.Write("Veuillez saisir un premier nombre : ");
-            int nombre1 = int.Parse(Console.ReadLine());
+            float nombre1 = float.Parse(Console.ReadLine());
             Console.WriteLine(nombre1);
             Console.Write("Veuillez saisir un second nombre : ");
-            int nombre2 = int.Parse(Console.ReadLine());
+            float nombre2 = float.Parse(Console.ReadLine());
             Console.WriteLine(nombre2);
             float moyenne = (nombre1 + nombre2) / 2;
             Console.WriteLine("La moyenne de " + nombre1 + " et " + nombre2 + " est : " + moyenne);
